# tests/unit/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ApplicationListItem": {
        "properties": {
          "application": {
            "$ref": "#/components/schemas/ApplicationRead"
          },
          "employer": {
            "$ref": "#/components/schemas/jobify_api__routes__schemas__EmployerRead"
          },
          "job": {
            "$ref": "#/components/schemas/JobRead"
          }
        },
        "required": [
          "application",
          "job",
          "employer"
        ],
        "title": "ApplicationListItem",
        "type": "object"
      },
      "ApplicationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApplicationListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ApplicationListResponse",
        "type": "object"
      },
      "ApplicationRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job_id",
          "status",
          "stage",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "ApplicationRead",
        "type": "object"
      },
      "ApplicationTimelineResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StageEventRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ApplicationTimelineResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobRead": {
        "properties": {
          "ctc_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Max"
          },
          "ctc_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctc Min"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "employer_verified": {
            "title": "Employer Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "title": "Locations",
            "type": "array"
          },
          "max_exp_years": {
            "title": "Max Exp Years",
            "type": "integer"
          },
          "min_exp_years": {
            "title": "Min Exp Years",
            "type": "integer"
          },
          "posted_at": {
            "format": "date-time",
            "title": "Posted At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "locations",
          "min_exp_years",
          "max_exp_years",
          "ctc_min",
          "ctc_max",
          "status",
          "posted_at",
          "employer_verified"
        ],
        "title": "JobRead",
        "type": "object"
      },
...
      "StageEventRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "from_stage": {
            "title": "From Stage",
            "type": "string"
          },
          "to_stage": {
            "title": "To Stage",
            "type": "string"
          }
        },
        "required": [
          "from_stage",
          "to_stage",
          "created_at"
        ],
        "title": "StageEventRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WithdrawRequest": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "WithdrawRequest",
        "type": "object"
      },
...
      "jobify_api__routes__schemas__EmployerRead": {
        "description": "Wire shape: a verified bool, not the underlying verified_at timestamp.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "verified"
        ],
        "title": "EmployerRead",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Jobify API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/applications": {
      "get": {
        "description": "Paginated list of the current applicant's applications (incl. withdrawn).\n\nCursor: base64 of {\"created_at\": ISO8601, \"application_id\": uuid}.\nOrder:  created_at DESC, id DESC.\nETag:   W/\"sha256(applicant_id|max_updated_at|count)\".",
        "operationId": "list_applications_v1_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Applications",
        "tags": [
          "applications"
        ]
      }
    },
    "/v1/applications/{application_id}": {
      "patch": {
        "description": "Withdraw an application.\n\nOnly ``applied \u2192 withdrawn`` is accepted.  Re-withdraw is a 200 no-op.\nAny other target status \u2192 400 invalid_transition.",
        "operationId": "patch_application_v1_applications__application_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Application",
        "tags": [
          "applications"
        ]
      }
    },
    "/v1/applications/{application_id}/resume": {
      "get": {
        "description": "Download the latest resume of the applicant who applied to one of the\nrecruiter's jobs.\n\nError ladder:\n- 401 \u2014 no/invalid bearer token (current_user).\n- 403 not_a_recruiter \u2014 caller is not a recruiter.\n- 404 not found \u2014 application doesn't exist, belongs to another employer's\n  job, applicant has no resume, or the application is soft-deleted.\n  Deliberately uniform to avoid existence leaks.\n\nEmits structured audit log ``recruiter.resume-accessed`` on success.",
        "operationId": "recruiter_download_application_resume_v1_applications__application_id__resume_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recruiter Download Application Resume",
        "tags": [
          "applications"
        ]
      }
    },
    "/v1/applications/{application_id}/timeline": {
      "get": {
        "description": "The applicant's stage journey. Owner-only (token identity), uniform\n404 across unknown-id and other-owner; actor identity never exposed.",
        "operationId": "get_application_timeline_v1_applications__application_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationTimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Application Timeline",
        "tags": [
          "applications"
        ]
      }
    },
...
    "/v1/jobs/{job_id}/apply": {
      "post": {
        "description": "Apply to an open job.\n\nIdempotent:\n- Existing applied row \u2192 200 with the existing row.\n- Existing withdrawn row \u2192 UPDATE back to applied, refresh created_at.\n- No existing row \u2192 INSERT. \u2192 201.\n\nError ladder: 401 (auth) \u2192 403 (role) \u2192 404 (job missing/closed/deleted).",
        "operationId": "apply_to_job_v1_jobs__job_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "feed",
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            },
            "description": "Already applied \u2014 the existing application is returned unchanged."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply To Job",
        "tags": [
          "applications"
        ]
      }
    },
...
  }
}